feat(apollo): add Divider
fixes #755
