+++ apps/apollo-stories/src/components/Divider/Divider.stories.tsx
@@ -1,0 +1,28 @@
+import { Divider } from "@axa-fr/design-system-apollo-react";
+import type { Meta, StoryObj } from "@storybook/react";
+
+const meta: Meta<typeof Divider> = {
+  title: "Components/Divider",
+  component: Divider,
+};
+
+export default meta;
+
+export const Default: StoryObj<typeof Divider> = {
+  name: "Divider",
+  render: (args) => (
+    <div>
+      <span>Hello</span>
+      <Divider {...args} /> <span>world!</span>
+    </div>
+  ),
+  args: {
+    classModifier: " ",
+  },
+  argTypes: {
+    classModifier: {
+      control: { type: "text" },
+      defaultValue: " ",
+    },
+  },
+};
