+++ client/apollo/css/src/Divider/DividerApollo.scss
@@ -1,0 +1,5 @@
+@use "./DividerCommon";
+
+.af-divider {
+  --divider-border-color: var(--axa-blue-20);
+}
