+++ client/apollo/css/src/Divider/DividerCommon.scss
@@ -1,0 +1,5 @@
+.af-divider {
+  margin: 0;
+  border: 0;
+  border-top: 1px solid var(--divider-border-color);
+}
