+++ client/apollo/css/src/Divider/DividerLF.scss
@@ -1,0 +1,5 @@
+@use "./DividerCommon";
+
+.af-divider {
+  --divider-border-color: var(--color-gray-300);
+}
