+++ client/apollo/css/src/apollo.scss
@@ -5,3 +5,4 @@
 @use "./Form/ItemMessage/ItemMessageApollo";
 @use "./Form/ItemLabel/ItemLabelApollo";
 @use "./Form/TextInput/TextInputApollo";
+@use "./Divider/DividerApollo";
