+++ client/apollo/css/src/look-and-feel.scss
@@ -5,3 +5,4 @@
 @use "./Form/ItemMessage/ItemMessageLF";
 @use "./Form/ItemLabel/ItemLabelLF";
 @use "./Form/TextInput/TextInputLF";
+@use "./Divider/DividerLF";
