+++ client/apollo/react/src/Divider/DividerApollo.tsx
@@ -1,0 +1,3 @@
+import "@axa-fr/design-system-apollo-css/dist/Divider/DividerApollo.scss";
+
+export { Divider } from "./DividerCommon";
