+++ client/apollo/react/src/Divider/DividerLF.tsx
@@ -1,0 +1,3 @@
+import "@axa-fr/design-system-apollo-css/dist/Divider/DividerLF.scss";
+
+export { Divider } from "./DividerCommon";
