+++ client/apollo/react/src/indexLF.ts
@@ -14,3 +14,4 @@ export { DebugGrid } from "./Grid/DebugGrid";
 export { ItemMessage } from "./Form/ItemMessage/ItemMessageLF";
 export { ItemLabel } from "./Form/ItemLabel/ItemLabelLF";
 export { TextInput } from "./Form/TextInput/TextInputLF";
+export { Divider } from "./Divider/DividerLF";
