--- client/look-and-feel/react/src/Divider/index.ts
@@ -1,3 +1,0 @@
-import "@axa-fr/design-system-look-and-feel-css/dist/Divider/Divider.scss";
-
-export { Divider } from "./Divider";
